Refactor ++
